pyproject.toml
[tool.poetry]
name = "casepro"
version = "1.3.29"
description = "Case management dashboard for UNICEF and partner organizations"
authors = ["Nyaruka <code@nyaruka.com>"]
license = "BSD-3"

[tool.poetry.dependencies]
python = "^3.6"
smartmin = "^2.3.8"
celery = "<5.0"
djangorestframework = "3.11.0"
phonenumbers = "^8.12.18"
gunicorn = "^20.0.4"
psycopg2-binary = "^2.8.6"
rapidpro-dash = "^1.5.6"
django-redis = "^4.12.1"
django-contrib-comments = "^2.0.0"
django-celery-email = "^3.0.0"
django-compressor = "^2.4"
django-el-pagination = "^3.3.0"
django-storages = "^1.11.1"
iso-639 = "^0.4.5"
colorama = "^0.4.4"
dj-database-url = "^0.5.0"
boto3 = "^1.17.11"
sentry-sdk = "^0.20.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
isort = "^5.7.0"
flake8 = "^3.8.4"
coverage = "^5.4"
polib = "^1.1.0"
codecov = "^2.1.11"
colorama = "^0.4.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
